launch a postgresql container if we can connect to the host docker
When the database connection parameters are not specified, the
image will try to spin up a postgresql container if it is able to
communicate with the host docker.

This is only possible if the following two options are specified in the
docker run command:
  `-v /var/run/docker.sock:/run/docker.sock`
  `-v $(which docker):/bin/docker`

This is primarily added to get the quick start guide to work without
much of a hassel.